Changed the Service query API. Hid database handler inside region classes.
Previously, there was confusing messy business to ensure that database
handlers and region classes would only be created once. Brainwave: those
classes should be Singletons! So... I made the db.DB class and region
subclasses Singletons. This simplifies things a lot.

Also, instead of separately managing the database handler for a region
(``dbh``), it is now encapsulated within the region class (a region "has a"
database handler). So, instead of passing around a ``region`` and a ``dbh``,
just pass around a ``region``, which has a handle on a database handler;
instead of creating a region and passing it to a DB, create a region
that creates its own handle to the singleton DB.

Part of all this was removing all the messy stuff from db.DB that managed
the regional tables and mappers globally. Everything that seemed like it
belonged more to "region" than "DB" was moved to the region base class.

Along with that, simplified the query API. Instead of passing ``region`` and
``dbh`` to query() and calling _beforeQuery(), just pass the region to the
service constructor. The query methods no longer accept the ``region`` and
``dbh`` args.

I think this turned out rather well, even though it's not what had planned to
do today.

All the other changes were related, changing call sites, etc.