* model/route.py * model/geocode.py * model/domain.py * model/address.py Added a __simplify__ method to all(?) domain objects. This methods returns a simplified representation of an object using dicts, list, and primitive types (just like JSON).
For many domain objects, this code was already in the __repr__ method, so it
was just moved from there and __repr__ now just does repr(__simplify__).

The name was inspired by the simplejson module, since the __simplify__ method
was added specifically to make it easier to get suitable input for
simplejson.dumps. (Before we were doing simplejson.dumps(eval(repr(obj)))).

* domain.py
  * (Edge): Removed clone()

* model/regions.py
Renamed _aliases to aliases