Moved rest.py from controllers directory to lib, out from user-land to framework-land (so to speak).
The way to use ``lib.rest`` is to replace 'from project.lib.base import *' with 'from project.lib.rest import *'
in controllers and then make the controller classes inherit from RestController instead of BaseController.

``lib.rest`` pulls in everything in ``lib.base`` so ``lib.rest`` is basically a drop in replacement that adds
REST support. Added a docstring in ``lib.rest`` explaining this.

Adjusted imports in affected files. Specifically, in ``regions`` and ``services`` controllers, replaced:
    from tripplanner.controllers.rest import RestController
with:
    from tripplanner.lib.rest import *