Used new-style Integrator to import milwaukeewi data. Modelled after portlandor. Made a few changes to make the base Integrator more generic.
Made code, permanent_id, and bikemode required fields for all regions.

Fixed bug in get_records that made it not return rows where all the
values are None, since this could very well be valid in some cases. It
might be good to add an any_not_none keyword arg.