Updated error handling throughout. This was first motivated by the desire to show better error messages to users. It ended up involving a somewhat major overhaul to the byCycle exception hierarchy.
The main change is that exceptions now have a concept of Title, Description,
and Explanation, where each item expounds on the previous--title is brief like
a heading, description is like a byline, and explanation is "prose."

The front end was updated to take advantage of this in consistent way.
Previously, the errors in the ``services`` controller used the byCycle
exceptions in a reasonable way, but the ``regions`` and ``routes`` controllers
didn't--they used a more ad hoc, text based solution. Now, when errors occur,
an ``exception`` is always added to the context, ``c``.

Files:
* core/trunk/byCycle/services/route.py
  * Added ``EmptyGraphError`` exception class

* core/trunk/byCycle/services/exceptions.py
  * Added ``title``, ``description``, and ``explanation`` attributes

* core/trunk/byCycle/services/geocode.py
  * Added ``MultipleAddressesNotFoundError`` exception class

* core/trunk/byCycle/services/normaddr.py
  * Added ``explanation`` when raising ``InputError`` when an address isn't
    "understood"