Fixed bicycle travel mode so that it takes into account MAX lines. In particular, if a street has a MAX line and a bike lane, we need to convert the street classification code to an equivalent non-MAX code (e.g., 5301 => 1300).
Related to this, don't set lengths for unknown street class codes in
_adjustRowForMatrix. Before, if the code didn't have certain values,
the length for the corresponding segment would be set to something big
so the segment wouldn't get used. Beside the fact that this was already
being handled in getEdgeWeight, getEdgeWeight is the *correct*
place to do stuff like that anyway.